YAHOO! I think the red-black tree implementation is working. It was working fine in "insert" mode, but crashing whilst trying to delete objects. Turns out to be a subtle issue regarding how one translates C pointers to fortran pointers.
There was one C statement that seems to have been the problem:

p = p->link[last] = single (q, dir)   # where q and p are pointers and last and dir are integers

I implemented in fortran as:
  pP => node_single_rotation_fn(pQ, iDir)
  pP%pLeft => pP

...assuming that since pP and pP%pLeft were being assigned the same "value", the ordering didn't matter. However, it appears that C evaluates the statement above in order from right to left! This first implementation doesn't flow the same way as the C implementation.

The correct implementation of this seems to be:
  pP%pLeft => node_single_rotation_fn(pQ, iDir)
  pP => pP%pLeft

Intuitively, this is first changing the address pointed to by the left pointer of pP, then making pP follow along. My original implementation clearly messes up all of the internal node connections.

Hopefully this is the last hiccup in implementing red-black.